+++ src/Cubes/getImage.cc
@@ -232,7 +232,7 @@ Either it has the wrong number of axes, or one axis has too large a range.\n");
 	  this->head.setIntFluxUnits();
 
 	  if(this->arrayAllocated){
-	    for(int i=0;i<this->numPixels;i++)
+	    for(size_t i=0;i<this->numPixels;i++)
 	      if(!this->isBlank(i)) this->array[i] = pow(scale * array[i] + offset, power);
 	  }
 	  if(this->par.isVerbose()) {
